[Create Account] Create the endpoint of accounts
This activity I built in the application layer the resource accounts. Right now it is simple because of not creating the service layer (I would like to use the model thinking about clean architecture, but at this moment I will do the simplest and evolve this project little by little.